refactor(web): extract a shared FolderExpandToggle
The workspace Files view's expand/collapse-all-folders toggle was inline in view-toolbar.tsx. Pull it into @/components/file-tree/folder-expand-toggle.tsx so the skills file tree can use the same affordance from one place. Behaviour for the workspace view is unchanged.
